third_party/WebKit/LayoutTests/fast/dom/Element/missing-arguments.html

<!doctype html>
<html>
<head>
<script src="../../../resources/js-test.js"></script>
</head>
<body>
<script>
var element = document.createElementNS("http://example.com/", "example");
[
    "getAttribute()",
    "getAttributeNS()",
    "getAttributeNS('http://example.com/')",
    "getElementsByClassName()",
    "getElementsByTagName()",
    "getElementsByTagNameNS()",
    "getElementsByTagNameNS('http://example.com/')",
    "hasAttributeNS()",
    "hasAttributeNS('http://example.com/')",
    "removeAttribute()",
    "removeAttributeNS()",
    "removeAttributeNS('http://example.com/')",
    "setAttribute()",
    "setAttribute('example')",
    "setAttributeNS()",
    "setAttributeNS('http://example.com/')",
    "setAttributeNS('http://example.com/', 'example')"
].forEach(function(expr)
{
    shouldThrow("element." + expr);
});
</script>
</body>
</html>
